Move implicit_hasher to its own module
